TjaTestingCompiler -> TestingCompiler
